+++ BACKEND/MOVIE/src/main/webapp/WEB-INF/views/common/header.jsp
@@ -95,9 +95,8 @@
                         <ul>
                            <li class="navBar"><a href="${contextPath}/movieList/detail_List">영화</a>
                               <ul class="menuBar" style="display: none">
-                                 <li><a href="${contextPath}/movieList/detail_List">상영중인
-                                       영화</a></li>
-                                 <li><a href="${contextPath}/movieList/all_List">전체 영화 목록</a></li>
+                                 <li><a href="${contextPath}/movieList/detail_List">상영 영화</a></li>
+                                 <li><a href="${contextPath}/movieList/all_List">전체 영화</a></li>
                               </ul>
                            </li>
                            <li class="navBar"><a href="${contextPath}/theater/normalTheater">상영관</a>
@@ -115,7 +114,7 @@
 
                            <li class="navBar"><a href="${contextPath}/eventList/detail_List">이벤트</a>
                               <ul class="menuBar" style="display: none">
-                                 <li><a href="${contextPath}/eventList/detail_List">진행중
+                                 <li><a href="${contextPath}/eventList/detail_List">진행
                                        이벤트</a></li>
                                  <li><a href="${contextPath}/eventList/end_List">지난 이벤트</a></li>
                               </ul>
@@ -123,7 +122,7 @@
 
                            <li class="navBar"><a href="${contextPath}/benefitsList/benefits_detail_List">혜택</a>
                               <ul class="menuBar" style="display: none">
-                                 <li><a href="${contextPath}/benefitsList/benefits_detail_List">진행중인 혜택</a></li>
+                                 <li><a href="${contextPath}/benefitsList/benefits_detail_List">진행 혜택</a></li>
                                  <li><a href="${contextPath}/benefitsList/benefits_end_List">지난 혜택</a></li>
                               </ul>
                            </li>
