+++ BACKEND/MOVIE/src/main/webapp/WEB-INF/views/movieList/detail_List.jsp
@@ -81,18 +81,18 @@
 																<span>${movie['movieTitle']}</span></a>
 															<div>
 																<p class="textSize">
-																	예매율&nbsp<span>${bookPercent[status.index]}</span>%
+																	예매율&nbsp<span><span class="revlikeVal">${bookPercent[status.index]}</span></span>%
 																</p>
 																<p class="textSize">
-																	개봉일&nbsp<span>${movie['movieOpen']}</span>
+																	개봉&nbsp<span>${movie['movieOpen']}</span>
 																</p>
 															</div>
 														</div>
 
 														<div>
 															<!-- 1좋아요버튼 -->
-															<button>${revLike[status.index]} /5.0</button>
-															<button>예매</button>
+															<button> <span class="revlikeVal">${revLike[status.index]}</span> / 5.0</button>
+															<a href="${contextPath}/reserve/choicePlay">예매</a>
 														</div>
 													</li>
 												</c:when>
