+++ BACKEND/MOVIE/src/main/webapp/WEB-INF/views/theater/normalTheater.jsp
@@ -34,138 +34,23 @@
 					<h2>전체극장</h2>
 					<div class="cont-wrap">
 						<div class="list_wrap">
+
+
+
+							
 							<ul class="area-wrap">
+								<c:forEach var="cinemaList" items="${cinemaList.cinemaList}">									
+								<li><span class="location" name="area">${cinemaList.cinemaArea}</span>
+									<div class="list-detail none"></div>
+								</li>
+							</c:forEach>
 
-
-								<li class="area-name"><span class="location yellowCol">서울</span>
-
-									<div class="list-detail block">
-										<ul>
-											<li><a href="https://www.naver.com">강남점 </a></li>
-											<li><a href="https://www.naver.com">건대입구점</a></li>
-											<li><a href="https://www.naver.com">구로점</a></li>
-											<li><a href="https://www.naver.com">노원점</a></li>
-											<li><a href="https://www.naver.com">동대문점</a></li>
-											<li><a href="https://www.naver.com">명동점</a></li>
-											<li><a href="https://www.naver.com">상봉점</a></li>
-											<li><a href="https://www.naver.com">상암점</a></li>
-											<li><a href="https://www.naver.com">신림점</a></li>
-											<li><a href="https://www.naver.com">신촌점</a></li>
-											<li><a href="https://www.naver.com">영등포점</a></li>
-											<li><a href="https://www.naver.com">왕십리점</a></li>
-											<li><a href="https://www.naver.com">용산점</a></li>
-											<li><a href="https://www.naver.com">잠실점</a></li>
-											<li><a href="https://www.naver.com">종로점</a></li>
-										</ul>
-									</div></li>
-
-
-								<li class="area-name"><span class="location">경기</span>
+								<!-- <li class="area-name"><span class="location" name="area">제주</span>
 									<div class="list-detail none">
 										<ul>
-											<li><a href="https://www.naver.com">광명점</a></li>
-											<li><a href="https://www.naver.com">구리점</a></li>
-											<li><a href="https://www.naver.com">김포점</a></li>
-											<li><a href="https://www.naver.com">남양주점</a></li>
-											<li><a href="https://www.naver.com">동탄점</a></li>
-											<li><a href="https://www.naver.com">부천점</a></li>
-											<li><a href="https://www.naver.com">성남점</a></li>
-											<li><a href="https://www.naver.com">수원점</a></li>
-											<li><a href="https://www.naver.com">안산점</a></li>
-											<li><a href="https://www.naver.com">안양점</a></li>
-											<li><a href="https://www.naver.com">부천점</a></li>
-											<li><a href="https://www.naver.com">송도점</a></li>
-											<li><a href="https://www.naver.com">주안점</a></li>
-											<li><a href="https://www.naver.com">일산점</a></li>
-											<li><a href="https://www.naver.com">하남점</a></li>
-
-
+											<li><a href=""> </a></li>
 										</ul>
-									</div></li>
-
-								<li class="area-name"><span class="location">충청</span>
-									<div class="list-detail none">
-										<ul>
-											<li><a href="https://www.naver.com">공주점</a></li>
-											<li><a href="https://www.naver.com">논산점</a></li>
-											<li><a href="https://www.naver.com">신탄진점</a></li>
-											<li><a href="https://www.naver.com">은행동점</a></li>
-											<li><a href="https://www.naver.com">충남대점</a></li>
-											<li><a href="https://www.naver.com">세종점</a></li>
-											<li><a href="https://www.naver.com">아산점</a></li>
-											<li><a href="https://www.naver.com">제천점</a></li>
-											<li><a href="https://www.naver.com">천안점</a></li>
-											<li><a href="https://www.naver.com">충주점</a></li>
-										</ul>
-									</div></li>
-								<li class="area-name"><span class="location">전라</span>
-									<div class="list-detail none">
-										<ul>
-											<li><a href="https://www.naver.com">상무점</a></li>
-											<li><a href="https://www.naver.com">전남대점</a></li>
-											<li><a href="https://www.naver.com">유스퀘어점</a></li>
-											<li><a href="https://www.naver.com">군산점</a></li>
-											<li><a href="https://www.naver.com">목포점</a></li>
-											<li><a href="https://www.naver.com">순천점</a></li>
-											<li><a href="https://www.naver.com">완주점</a></li>
-											<li><a href="https://www.naver.com">익산점</a></li>
-											<li><a href="https://www.naver.com">장성점</a></li>
-											<li><a href="https://www.naver.com">전주점</a></li>
-
-										</ul>
-									</div></li>
-								<li class="area-name"><span class="location">경남</span>
-									<div class="list-detail none">
-										<ul>
-											<li><a href="https://www.naver.com">김해점</a></li>
-											<li><a href="https://www.naver.com">마산점</a></li>
-											<li><a href="https://www.naver.com">부산대점</a></li>
-											<li><a href="https://www.naver.com">서면점</a></li>
-											<li><a href="https://www.naver.com">센텀시티점</a></li>
-											<li><a href="https://www.naver.com">해운대점</a></li>
-											<li><a href="https://www.naver.com">울산역점</a></li>
-											<li><a href="https://www.naver.com">태화강점</a></li>
-											<li><a href="https://www.naver.com">진주점</a></li>
-											<li><a href="https://www.naver.com">창원점</a></li>
-
-
-										</ul>
-									</div></li>
-								<li class="area-name"><span class="location">경북</span>
-									<div class="list-detail none">
-										<ul>
-											<li><a href="https://www.naver.com">경주점</a></li>
-											<li><a href="https://www.naver.com">구미점</a></li>
-											<li><a href="https://www.naver.com">김천점</a></li>
-											<li><a href="https://www.naver.com">경북대점</a></li>
-											<li><a href="https://www.naver.com">동대구점</a></li>
-											<li><a href="https://www.naver.com">동성로점</a></li>
-											<li><a href="https://www.naver.com">안동점</a></li>
-											<li><a href="https://www.naver.com">영주점</a></li>
-											<li><a href="https://www.naver.com">영천점</a></li>
-											<li><a href="https://www.naver.com">포항점</a></li>
-
-										</ul>
-									</div></li>
-								<li class="area-name"><span class="location">강원</span>
-									<div class="list-detail none">
-										<ul>
-											<li><a href="https://www.naver.com">강릉점</a></li>
-											<li><a href="https://www.naver.com">동해점</a></li>
-											<li><a href="https://www.naver.com">속초점</a></li>
-											<li><a href="https://www.naver.com">원주점</a></li>
-											<li><a href="https://www.naver.com">춘천점</a></li>
-
-										</ul>
-									</div></li>
-								<li class="area-name"><span class="location">제주</span>
-									<div class="list-detail none">
-										<ul>
-											<li><a href="https://www.naver.com">제주점ㅋ</a></li>
-
-
-										</ul>
-									</div></li>
+									</div></li> -->
 
 							</ul>
 						</div>
@@ -235,7 +120,6 @@
 	<jsp:include page="/WEB-INF/views/common/footer.jsp" />
 
 	
-	<script src="${contextPath}/resources/js/common/header.js"></script>
 	<script src="${contextPath}/resources/js/theater/theater.js"></script>
 </body>
