+++ BACKEND/MOVIE/src/main/webapp/resources/css/helpDesk/helpDesk_Home.css
@@ -14,11 +14,11 @@ main>section {
 
 
 .aslide-list>li:nth-child(1)>a {
-    background-color: rgb(73, 73, 73);
+    background-color: #FFC400;
 }
 
 .aslide-list>li:nth-child(1)>a {
-    color: rgb(251, 251, 251);
+    color: black;
 }
 
 h2 {
