+++ BACKEND/MOVIE/src/main/webapp/resources/css/helpDesk/lost_List.css
@@ -15,46 +15,55 @@ main>section {
 }
 
 #lostArea {
-    background-color: rgba(73, 73, 73);
+    background-color: #191919;
+
     color: white;
     margin-right:5px;
     border:none;
     outline:none;
+    border-radius: 5px;
 }
 
 #lostName {
-    background-color: rgba(73, 73, 73);
+    background-color: #191919;
+
     color:white;
     margin-right:5px;
     border:none;
     outline:none;
+    border-radius: 5px;
 }
 
 .search-area button{
-    background-color: rgba(73, 73, 73);
-    margin-left:5px;
+    background-color: #191919;
+
     color:white;
     height:100%;
     outline: none;
     border:none;
     padding:0 10px;
     cursor: pointer;
+    border-top-right-radius: 5px;
+    border-bottom-right-radius: 5px;
 }
 
 .search-area input{
-    background-color: rgba(73, 73, 73);
+    background-color: #191919;
+
     color:white;
     border:none;
     outline:none;
     padding-left:10px;
+    border-top-left-radius: 5px;
+    border-bottom-left-radius: 5px;
 }
 
 .aslide-list>li:nth-child(5)>a {
-    background-color: rgb(73, 73, 73);
+    background-color: #FFC400;
 }
 
 .aslide-list>li:nth-child(5)>a {
-    color: rgb(251, 251, 251);
+    color: black;
 }
 
 .notice-wraaper>div:nth-child(2) {
@@ -88,6 +97,7 @@ h2 {
 
 .search-area>p {
     line-height: 28px;
+    padding-left:10px;
 }
 
 .search-area>p>span {
@@ -124,59 +134,57 @@ h2 {
 }
 
 .table-wrap {
-    width: 784px;
-    
+	width: 784px;
 }
 
 table {
-    width: 100%;
-    border-collapse: collapse;
+	width: 100%;
+	border-collapse: collapse;
 }
 
 thead * {
-    padding:10px;   border-bottom: 2px solid rgb(73, 73, 73);
+	padding: 10px;
+	background-color: rgb(52,52,52);;
 }
 
-
 thead th:nth-child(1) {
-    width: 25%;
-    height:50px;
+	width: 15%;
+	height: 50px;
 }
 
 thead th:nth-child(2) {
-    width: 45%;
+	width: 65%;
 }
- 
+
 thead th:nth-child(3) {
-    width: 15%;
+	width: 15%;
 }
+
 thead th:nth-child(4) {
-    width: 15%;
+	width: 15%;
 }
 
-
-tbody{
-    text-align: center;
+tbody {
+	text-align: center;
 }
 
-.row *{
-    padding:10px;
-    height:40px;
-    border-bottom: 1px solid rgba(73, 73, 73, 0.4);
+.row * {
+	padding: 10px;
+	height: 50px;
 }
 
 .row td:nth-child(2) {
-    text-align: left;
-    max-width:450px;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    overflow: hidden;
-    padding:10px
+	text-align: left;
+	max-width: 450px;
+	text-overflow: ellipsis;
+	white-space: nowrap;
+	overflow: hidden;
+	padding: 10px
 }
 
-.row td:nth-child(2) > a:hover{
-    text-decoration: underline;
-    color:rgb(251, 251, 251);
+.row td:nth-child(2)>a:hover {
+	text-decoration: underline;
+	color: rgb(251, 251, 251);
 }
 
 
@@ -262,3 +270,11 @@ tbody{
 	font-size:13px;
 }
 
+
+.row:nth-child(even) {
+    background-color: rgb(52,52,52);
+}
+
+.selected_Cp {
+    color: rgb(255, 196, 0);
+}
