+++ BACKEND/MOVIE/src/main/webapp/resources/css/helpDesk/lost_detail.css
@@ -252,9 +252,9 @@ a.hover {
 }
 
 .aslide-list>li:nth-child(5)>a {
-    background-color: rgb(73, 73, 73);
+    background-color: #FFC400;
 }
 
 .aslide-list>li:nth-child(5)>a {
-    color: rgb(251, 251, 251);
+    color: black;
 }
