+++ BACKEND/MOVIE/src/main/webapp/resources/css/helpDesk/lost_form.css
@@ -9,11 +9,11 @@ main>section {
 }
 
 .aslide-list>li:nth-child(5)>a {
-    background-color: rgb(73, 73, 73);
+    background-color: #FFC400;
 }
- 
+
 .aslide-list>li:nth-child(5)>a {
-    color: rgb(251, 251, 251);
+    color: black;
 }
 .notice-wraaper>div:nth-child(2) {
     width: 82%;
