+++ BACKEND/MOVIE/src/main/webapp/resources/css/helpDesk/mTm_List.css
@@ -9,11 +9,11 @@ main>section {
 }
 
 .aslide-list>li:nth-child(3)>a {
-	background-color: rgb(73, 73, 73);
+    background-color: #FFC400;
 }
 
 .aslide-list>li:nth-child(3)>a {
-	color: rgb(251, 251, 251);
+    color: black;
 }
 
 .notice-wraaper>div:nth-child(2) {
@@ -243,8 +243,8 @@ tbody {
 }
 
 .qlrhdro:hover {
-	color: red !important;
-	text-decoration: none !important;
+	color:white;
+	text-decoration: underline !important;
 }
 
 .replyComplete{
