+++ BACKEND/MOVIE/src/main/webapp/resources/css/helpDesk/mTm_detail.css
@@ -76,11 +76,11 @@ h2 {
 }
 
 .aslide-list>li:nth-child(3)>a {
-    background-color: rgb(73, 73, 73);
+    background-color: #FFC400;
 }
 
 .aslide-list>li:nth-child(3)>a {
-    color: rgb(251, 251, 251);
+    color: black;
 }
