+++ BACKEND/MOVIE/src/main/webapp/resources/css/helpDesk/notice_detail.css
@@ -123,9 +123,9 @@ h2 {
 }
 
 .aslide-list>li:nth-child(2)>a {
-    background-color: rgb(73, 73, 73);
+    background-color: #FFC400;
 }
 
 .aslide-list>li:nth-child(2)>a {
-    color: rgb(251, 251, 251);
+    color: black;
 }
