+++ BACKEND/MOVIE/src/main/webapp/resources/css/theater/theater.css
@@ -15,9 +15,8 @@ h2 {
 }
 
 .cont-wrap{
-    border:3px solid rgba(255, 255, 255, 0.3);
+    border:3px solid#FCC400;
     border-radius: 10px;
-    padding:16px 0;
 }
 
 .list_wrap {
@@ -30,25 +29,38 @@ h2 {
 .area-wrap {
     display: flex;
     border-radius: 10px;
-}
-
-
-.area-name >span{
-    padding:20px 40px;
-    z-index: 10;
+    height:64px;
 }
 
 
 
-.area-wrap>li {
+.area-wrap li {
     width: 10%;
     cursor: pointer;
     text-align: center;
     font-size: 1.2rem;
+    line-height: 62px;
+
 }
 
+.area-wrap li:nth-child(1){
+    background-color: #FCC400;
+}
+
+.area-wrap li:nth-child(1) > a > span{
+    color:black;
+}
+
+.area-name >span{
+    padding:20px 40px;
+    z-index: 10;
+    color:black;
+}
+
+
+
 .list-detail {
-    margin-top:5px;
+    margin-top:20px;
     display: none;
     width: 1040px;
     height: 100%;
@@ -68,6 +80,7 @@ h2 {
     height: 30px;
     display: flex;
     z-index: 99;
+
     
 }
 
@@ -95,12 +108,18 @@ h2 {
 
 .yellowCol{
     color:black;
-    background-color: rgba(210,210,210, 0.8);
+    font-weight: bold;
+    background-color:  #FFC400;
+}
+
+.location{
+    font-weight: bold;
 }
 
 .area-wrap > .area-name:nth-child(1) > .location{
     border-top-left-radius: 8px;
 }
+
 
 .event-wrap{
     width:1040px;
