+++ BACKEND/MOVIE/src/main/webapp/resources/js/theater/theater.js
@@ -1,19 +1,27 @@
+$(document).ready(function () {
+  $(".area-name").click(function () {
+    var index = $(this).index();
 
+    $(".list-detail").addClass("none");
 
-$(document).ready(function() {
-    $(".area-name").click(function() {
-        var index = $(this).index();
+    $(this)
+      .find(".list-detail")
+      .removeClass("none")
+      .css("left", -(index * 97.51) + "px");
+  });
 
-        $(".list-detail").addClass("none");
-
-        $(this).find(".list-detail").removeClass("none").css("left", -(index * 97.51) + "px");
-    });
-});
-
-$(document).ready(function() {
-    $('.location').click(function() {
-      $('.location').removeClass('yellowCol');
-      
-      $(this).addClass('yellowCol');
+  $(".location").click(function () {
+    var area = $(this).text();
+    console.log(area);
+    $.ajax({
+      url: "/movie/theater/normalTheater/selectArea",
+      type: "GET",
+      data: { area: area },
+      success: function (response) {
+        console.log(response);
+        $(".list-detail").html();
+      },
+      error: function () {},
     });
   });
+});
